requirements.txt
gradio
transformers
ollama
openai
python-dotenv
huggingface_hub
hf-cli